Sat, Dec 2, 2023, 8:12 PM +00:00

+++ pset5/speller/dictionary.c
@@ -44,7 +44,7 @@ bool check(const char *word)
 // Hashes word to a number
 unsigned int hash(const char *word)
 {
-    unsigned int long hash = 5381;
+    unsigned int hash = 5381;
     int c;
 
     while((c = *word++))
